+++ Sources/TuistGraph/Models/CarthageDependencies.swift
@@ -6,14 +6,18 @@ public struct CarthageDependencies: Equatable {
     public let dependencies: [Dependency]
     /// List of platforms for which you want to install depedencies.
     public let platforms: Set<Platform>
-    /// Indicates whether Carthage produces XCFrameworks or regular frameworks.
-    public let useXCFrameworks: Bool
+    /// List of options for Carthage installation.
+    public let options: Set<Options>
 
     /// Initializes the carthage dependency with its attributes.
-    public init(_ dependencies: [Dependency], platforms: Set<Platform>, useXCFrameworks: Bool) {
+    public init(
+        _ dependencies: [Dependency],
+        platforms: Set<Platform>,
+        options: Set<Options>
+    ) {
         self.dependencies = dependencies
         self.platforms = platforms
-        self.useXCFrameworks = useXCFrameworks
+        self.options = options
     }
 
     /// Returns `Cartfile` representation.
@@ -65,5 +69,10 @@ public extension CarthageDependencies {
                 return #""\#(revision)""#
             }
         }
+    }
+
+    enum Options: Equatable {
+        case useXCFrameworks
+        case noUseBinaries
     }
 }